new elf program
This patch introduces a thread method that allows a new program to be created for
the calling thread. The new program is created at a different slice of the address space.
This effectively imposes a lower amount of memory each program can use. But the limit is
still big enough, and should be enough for anybody (tm).

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>